Fix compiler warning : comparison between signed and unsigned int

+++ src/gui-sdl/dlgAbout.c
@@ -46,7 +46,7 @@ static SGOBJ aboutdlg[] =
  */
 void Dialog_AboutDlg(void)
 {
-	if (strlen(aboutstr) > aboutdlg[0].w)
+	if ((int)strlen(aboutstr) > aboutdlg[0].w)
 	{
 		/* Shorten the name if it is too long */
 		char *p = strrchr(aboutstr, '(');
